Controller and AI Update
Controller Updated
The ‘ [ ‘ / ‘ ] ‘ keys cause the camera to rotate left and right
instead of the left/right arrow keys

Protect AI updated
Seems the AI script was changed to an old version during git commit. AI
triggers rag doll in a simple way. When the AI is in attack his
collider is tagged as an ‘Obstacle’ then the OnCollisionEnter Routine
in KayaLoco handles the call to goForRagdoll