Jump Effects and Ragdoll Debug
Moving jump to update without proper conditions allowed for jump to be
played in idle state. This was fixed by checking to make sure the
player is moving and the kaya AC is not in transition

When transition to rag doll while moving..blood particles and camera
would drift. This was due to the low friction physics material sliding
like ice on the floor after disabling the AC. Changing physics material
back to none when initiating rag doll fixed this.

For some reason this update broke the Protect AI. Will not chase.
Turret AI is still working fine….not sure why?