Swing Platform Mock-up
Swinging platform basics completed. Character can jump on and off
platform.
Character will move on the platform since his capsule collider has zero
friction. This can be improved by scripting collision detection with
platform, and this is why platform moves so slow

Character is not made a parent of the platform at anytime so moving
platform problems not a big deal.